Be more consistent
- Check for Float, not just Numeric
- Say a_kind_of instead of be_a_kind_of

+++ spec/forms/idv/api_image_upload_form_spec.rb
@@ -136,7 +136,7 @@ RSpec.describe Idv::ApiImageUploadForm do
           state_id_type: 'drivers_license',
           success: true,
           user_id: document_capture_session.user.uuid,
-          vendor_request_time_in_ms: a_kind_of(Numeric),
+          vendor_request_time_in_ms: a_kind_of(Float),
         )
       end
